# 順列最適化問題（Permutation Optimization Problem）& modInverse Dev/macbook pro