--- a/Algorithm/BinarySearch/leetcode/4. Median of Two Sorted Arrays/MedianofTwoSortedArrays.ts
+++ b/Algorithm/BinarySearch/leetcode/4. Median of Two Sorted Arrays/MedianofTwoSortedArrays.ts
@@ -1,0 +1,35 @@
+function findMedianSortedArrays(nums1: number[], nums2: number[]): number {
+    if (nums1.length > nums2.length) {
+        [nums1, nums2] = [nums2, nums1]; // nums1 を常に短い方に
+    }
+
+    const m = nums1.length;
+    const n = nums2.length;
+    let low = 0,
+        high = m;
+
+    while (low <= high) {
+        const partitionX = Math.floor((low + high) / 2);
+        const partitionY = Math.floor((m + n + 1) / 2) - partitionX;
+
+        const maxLeftX = partitionX === 0 ? -Infinity : nums1[partitionX - 1];
+        const minRightX = partitionX === m ? Infinity : nums1[partitionX];
+
+        const maxLeftY = partitionY === 0 ? -Infinity : nums2[partitionY - 1];
+        const minRightY = partitionY === n ? Infinity : nums2[partitionY];
+
+        if (maxLeftX <= minRightY && maxLeftY <= minRightX) {
+            if ((m + n) % 2 === 0) {
+                return (Math.max(maxLeftX, maxLeftY) + Math.min(minRightX, minRightY)) / 2;
+            } else {
+                return Math.max(maxLeftX, maxLeftY);
+            }
+        } else if (maxLeftX > minRightY) {
+            high = partitionX - 1; // 左を減らす
+        } else {
+            low = partitionX + 1; // 左を増やす
+        }
+    }
+
+    throw new Error('Input arrays are not sorted properly.');
+}