提出日時 | 問題 | ユーザ | 言語 | 得点 | コード長 | 結果 | 実行時間 | メモリ |  
-- | -- | -- | -- | -- | -- | -- | -- | -- | --
2025-06-25 18:06:47 | A71 - Homework | myoshizumi | PHP (php 8.2.8) | 1000 | 450 Byte | AC | 17 ms | 21576 KiB | 詳細
2025-06-25 18:04:10 | A71 - Homework | myoshizumi | Go (go 1.20.6) | 1000 | 842 Byte | AC | 1 ms | 1636 KiB | 詳細
2025-06-25 18:01:44 | A71 - Homework | myoshizumi | Python (CPython 3.11.4) | 1000 | 409 Byte | AC | 12 ms | 8612 KiB | 詳細
2025-06-25 17:55:06 | A71 - Homework | myoshizumi | TypeScript 5.1 (Node.js 18.16.1) | 1000 | 672 Byte | AC | 40 ms | 42900 KiB | 詳細
2025-06-25 17:48:22 | A71 - Homework | myoshizumi | JavaScript (Node.js 18.16.1) | 1000 | 636 Byte | AC | 39 ms | 42804 KiB | 詳細